+++ lib/phoenix/sync/sandbox/expiry_manager.ex
@@ -1,0 +1,13 @@
+if Phoenix.Sync.sandbox_enabled?() do
+  defmodule Phoenix.Sync.Sandbox.ExpiryManager do
+    use GenServer
+
+    def start_link(args) do
+      GenServer.start_link(__MODULE__, args, name: Electric.ShapeCache.ExpiryManager.name(args))
+    end
+
+    def init(_) do
+      {:ok, []}
+    end
+  end
+end